+++ lib/src/vector/float32x4/float32x4_data_store_mixin.dart
@@ -2,8 +2,8 @@ import 'dart:typed_data';
 
 import 'package:ml_linalg/src/vector/ml_vector_data_store.dart';
 
-class Float32x4DataStoreMixin
-    implements MLVectorDataStore<Float32x4, Float32x4List> {
+mixin Float32x4DataStoreMixin
+    implements VectorDataStore<Float32x4, Float32x4List> {
   @override
   Float32x4List data;