+++ lib/src/vector/float32x4/float32x4_vector.dart
@@ -22,24 +22,19 @@ import 'package:ml_linalg/vector.dart';
 /// - Each SIMD-typed value is a "cell", that contains several floating point values (2 or 4).
 /// - Sequence of SIMD-values forms a "computation lane", where computations are performed with each floating point element
 /// simultaneously (in parallel)
-class Float32x4Vector extends Object
-    with
+class Float32x4Vector with
         IterableMixin<double>,
         Float32ListFactoryMixin,
         Float32x4OperationsMixin,
         Float32x4VectorFactoryMixin,
         Float32x4DataStoreMixin,
-        MLSimdVectorFastIterableMixin<Float32x4, Float32x4List>,
-        MLSimdVectorOperationsMixin<Float32x4, Float32x4List>
-    implements MLVector {
-  @override
-  final bool isMutable;
-
+        SimdVectorFastIterableMixin<Float32x4, Float32x4List>,
+        SimdVectorOperationsMixin<Float32x4, Float32x4List>
+    implements Vector {
   /// Creates a vector from collection
   Float32x4Vector.from(Iterable<double> source, {this.isMutable = false}) {
     length = source.length;
-    final List<double> _source =
-        source is List ? source : source.toList(growable: false);
+    final _source = source.toList(growable: false);
     data = convertCollectionToSIMDList(_source);
   }
 
@@ -71,4 +66,7 @@ class Float32x4Vector extends Object
     final source = List<double>.generate(length, (_) => random.nextDouble());
     data = convertCollectionToSIMDList(source);
   }
+
+  @override
+  final bool isMutable;
 }