+++ lib/src/vector/ml_vector_data_store.dart
@@ -1,4 +1,4 @@
-abstract class MLVectorDataStore<E, S extends List<E>> {
+abstract class VectorDataStore<E, S extends List<E>> {
   S data;
   int length;
   List<double> toList({bool growable = false});